Create DBConnection
Login info and location of mysql Database is stored in the DBInfo class. 
Creating the connection to the Database is done in the DBConnection class.
Verifying if the username and password are valid is done in the checkCredentials class.
Inserting the username and password, when a user registers for an account, is done in the insertUser class.